Initial Commit of dotfiles.
Just spent the last day or so reinstalling linux due to circular
dependencies that I couldn't figure out. Also, because I was 3
years out of date, and the only way to fix that was a clean install.
Of course I forgot to save my dotfiles. Now they're here, so that
should never happend again. Hopefully.